requirements.txt
click==6.6
Flask==0.11.1
Flask-Cors==3.0.2
fonttools==3.1.2
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.12
six==1.10.0
Werkzeug==0.11.11
